+++ Cargo.toml
@@ -30,9 +30,6 @@ version = "0.10.1"
 package = "image"
 version = "0.24"
 
-[dev-dependencies]
-mockall = "0.11.3"
-
 # compile all external dependencies in release mode
 [profile.dev.package."*"]
 opt-level = 3
