+++ audio/Cargo.toml
@@ -31,16 +31,6 @@ rb = "0.3.2"
 arrayvec = "0.7.1"
 rubato = "0.12.0"
 
-[target.'cfg(target_os = "windows")'.dependencies]
-once_cell = "1.17.1"
-
-[target.'cfg(target_os = "windows")'.dependencies.windows]
-version = "0.48.0"
-features = [
-    "Win32_Foundation",
-    "Win32_UI_WindowsAndMessaging"
-]
-
 [dev-dependencies]
 mockall = "0.11.3"
