+++ audio/src/lib.rs
@@ -1,6 +1,6 @@
 #![warn(rust_2018_idioms)]
 #![deny(missing_debug_implementations)]
-#![deny(unsafe_code)]
+#![forbid(unsafe_code)]
 
 pub mod metadata;
 pub mod player;
