+++ shared/Cargo.toml
@@ -3,3 +3,16 @@ name = "clef_shared"
 version = "0.0.1"
 edition = "2021"
 authors = [ "Dan Knutson <dan.knutson@gmail.com>" ]
+
+[dependencies]
+log = { version = "0.4", features = ["release_max_level_info"] }
+
+[target.'cfg(target_os = "windows")'.dependencies]
+once_cell = "1.17.1"
+
+[target.'cfg(target_os = "windows")'.dependencies.windows]
+version = "0.48.0"
+features = [
+    "Win32_Foundation",
+    "Win32_UI_WindowsAndMessaging"
+]
