+++ src/test_util.rs
@@ -2,8 +2,9 @@ use std::str::FromStr;
 
 use camino::Utf8PathBuf;
 
+use clef_db::queries::*;
+
 use crate::app::crawler::CrawledAlbum;
-use crate::db::queries::*;
 
 pub fn fake_album() -> CrawledAlbum {
     let album_id = AlbumId::new(1);
